Adding Splitting Functionalities

+++ ExpensesManager/src/main/java/com/tools/expenseManager/dao/ContributorMetadataInput.java
@@ -9,6 +9,6 @@ public class ContributorMetadataInput {
     private String accountNumber;  // Account number of the Contributor
     private boolean expenseCreator;  // it is the value for person who created the expense
     private boolean expenseContributor; // check the person contributes in the expense
-    private int contributorShareExpense;
+    private float contributorShareExpense;
     private FetchExpense expense;
 }
