+++ ExpensesManager/src/main/java/com/tools/expenseManager/dao/ExpenseMetadataInput.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class ExpenseMetadataInput {
     private  String expenseName;
     private  String description;
-    private  int amount;
+    private  float amount;
     private  boolean expenseSplitCheck; //if its true then count the number of contributors and divide the splitting amount as per they paid and update the amount in contributor splitAmountPortion field
     private String paymentMode; // Check the final payment made by UPI,CASH,MANDATE
     private String expensePaymentStatus; // True if payment for expense is completed or False if not completed
